requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dominate==2.8.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
fonttools==4.44.0
fsspec==2023.10.0
fugashi==1.3.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.19.4
idna==3.4
install==1.3.5
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
jsonfy==0.4
kiwisolver==1.4.5
loguru==0.7.2
manga-ocr==0.1.11
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sympy==1.12
termcolor==2.4.0
thop==0.1.1.post2209072238
tokenizers==0.15.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.36.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.207
unidic-lite==1.0.8
urllib3==2.0.7
visitor==0.1.3
Werkzeug==3.0.1
